Compiles after big refactor.
This is mostly refactoring Exercise so that sets, expected, and current are all
bundled together into exercise info classes. As part of this Exercise instances
are now replicated into workouts. This is basically de-normalizing the data to
make it quite a bit easier to use but a bit harder to modify (since changes have
to be replicated through the workouts).